improve static provider design
