+++ package.json
@@ -36,6 +36,7 @@
     "source-map-loader": "^4.0.1",
     "ts-loader": "^9.4.2",
     "unleash-client": "^3.19.0",
+    "webpack": "^5.88.2",
     "winston": "^3.8.2",
     "winston-loki": "^6.0.6"
   },
@@ -53,6 +54,7 @@
     "eslint-config-prettier": "^8.6.0",
     "eslint-plugin-prettier": "^4.2.1",
     "mocha": "^10.2.0",
+    "mocha-each": "^2.0.1",
     "msw": "^1.2.1",
     "nyc": "^15.1.0",
     "prettier": "2.8.3",
