+++ src/providers/interval.ts
@@ -1,0 +1,21 @@
+class Interval {
+  #callback: () => void;
+  #interval: NodeJS.Timeout;
+  #seconds: number;
+
+  constructor(seconds: number, callback: () => void) {
+    this.#callback = callback;
+    this.#seconds = seconds;
+  }
+
+  start() {
+    this.#callback();
+    this.#interval = setInterval(this.#callback, this.#seconds * 1000);
+  }
+
+  stop() {
+    clearInterval(this.#interval);
+  }
+}
+
+export default Interval;
