+++ src/providers/realtime-providers/dataAdapter.ts
