+++ src/providers/realtime-providers/vehiclePosition/groupVehicles.ts
