+++ src/providers/realtime-providers/vehiclePosition/vehicle.ts
