+++ src/providers/static-providers/gtfs/parsers/parser.d.ts
@@ -1,0 +1,3 @@
+export interface Parser<T_in, T_out> {
+  parseRow(data: T_in): T_out;
+}
