+++ src/providers/static-providers/index.ts
@@ -1,0 +1,1 @@
+// TODO: abstract static-providers to support non-GTFS data sources
